# Conflicts:
#	_maps/map_files/Delta/delta.dmm
#	_maps/map_files/MetaStation/MetaStation.v41A.II.dmm
#	_maps/map_files/MetaStation/z4.dmm
#	_maps/map_files/RandomRuins/SpaceRuins/derelict5.dmm
#	_maps/map_files/RandomRuins/SpaceRuins/spacebar.dmm
#	_maps/map_files/RandomZLevels/spacebattle.dmm
#	_maps/map_files/cyberiad/cyberiad.dmm
#	_maps/map_files/cyberiad/z4.dmm
#	_maps/map_files/cyberiad/z6.dmm
#	_maps/map_files/generic/z5.dmm
#	code/__DEFINES/misc.dm
#	code/__HELPERS/unsorted.dm
#	code/game/objects/structures/lattice.dm
#	code/game/turfs/space/space.dm
#	code/modules/mob/mob.dm
#	code/modules/ruins/ruin_areas.dm
#	paradise.dme

+++ code/modules/mob/mob_grab.dm
@@ -387,15 +387,24 @@
 	if(M == assailant && state >= GRAB_AGGRESSIVE) //no eatin unless you have an agressive grab
 		if(checkvalid(user, affecting)) //wut
 			var/mob/living/carbon/attacker = user
+			
+			if(affecting.buckled)
+				to_chat(user, "<span class='warning'>[affecting] is buckled!</span>")
+				return
+
 			user.visible_message("<span class='danger'>[user] is attempting to devour \the [affecting]!</span>")
 
 			if(!do_after(user, checktime(user, affecting), target = affecting)) return
+
+			if(affecting.buckled)
+				to_chat(user, "<span class='warning'>[affecting] is buckled!</span>")
+				return
 
 			user.visible_message("<span class='danger'>[user] devours \the [affecting]!</span>")
 			if(affecting.mind)
 				add_attack_logs(attacker, affecting, "Devoured")
 
-			affecting.loc = user
+			affecting.forceMove(user)
 			attacker.stomach_contents.Add(affecting)
 			qdel(src)
